FDI Refactor
move common base api view into lib/common app
